Tests/LibWeb/Text/expected/svg-relayout-inline-boundary-structural.txt

svg=40,20
partial-layouts=0 full-layouts=1
FAIL
